Change name of contract to MultiFlux and move unit tests into SnappBase test file
