+++ contracts/libraries/wrapppers/MultiFlux.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.5.0;
 import "../../SnappBase.sol";
 
 
-contract MultiRequest is SnappBase {
+contract MultiFlux is SnappBase {
     // Testing the filled deposit batch via deposit() function
     function multiDeposit(uint8 tokenId, uint128 amount, uint multiplier) public {
         for (uint8 i = 0; i < multiplier; i++) {
